refactor: audit sécurité/optimisation (suite) - montants, contraintes SQL, perf
- Montants stockés en centimes (int) au lieu de double, partout (DB, DAOs,
  écrans, export CSV, sync Supabase) — élimine les erreurs d'arrondi
  flottant sur les calculs financiers. Migration v6->v7 testée.
- Activation réelle de PRAGMA foreign_keys (les .references() du schéma
  étaient décoratives jusqu'ici), ON DELETE CASCADE (Cours->Eleves) et
  SET NULL (Cours->Tarifs), contraintes CHECK (prix>0, durée>0, jour 1-7,
  montant>=0), index unique anti double-booking. Migration v7->v8 testée
  avec nettoyage de doublons existants.
- Documentation des 3 patterns de soft-delete distincts (actif/statut/dateFin)
  pour éviter une confusion future.
- Délai avant ré-authentification de l'app : 5 -> 15 minutes.
- Mutualisation du tri/groupement par jour hebdo (eleves_screen et
  detail_mois_screen dupliquaient le même algorithme).
- Suppression du N+1 sur eleveParIdProvider (lookup synchrone via
  elevesMapProvider au lieu d'une requête par carte affichée).
- Récap mensuel des paiements calculé par agrégation SQL (GROUP BY) au
  lieu de charger tout l'historique des cours en mémoire côté Dart.

+++ lib/shared/models/mois_info.dart
@@ -1,8 +1,9 @@
 class MoisInfo {
   final DateTime mois;
   final int nbCours;
-  final double montantTotal;
-  final double montantPaye;
+  // Montants en centimes.
+  final int montantTotal;
+  final int montantPaye;
 
   const MoisInfo({
     required this.mois,
@@ -12,5 +13,5 @@ class MoisInfo {
   });
 
   bool get toutPaye => nbCours > 0 && montantPaye >= montantTotal;
-  double get montantRestant => montantTotal - montantPaye;
+  int get montantRestant => montantTotal - montantPaye;
 }
